Renaming KitCheckoutProvider to SequenceCheckoutProvider
