+++ packages/checkout/src/components/CopyButton.tsx
