+++ packages/checkout/src/components/DefaultIcon.tsx
