+++ packages/checkout/src/components/NavigationHeader.tsx
@@ -1,8 +1,8 @@
 import { IconButton, ChevronLeftIcon, Text, ModalPrimitive } from '@0xsequence/design-system'
 import React from 'react'
 
-import { HEADER_HEIGHT } from '../../constants'
-import { useNavigation } from '../../hooks/useNavigation'
+import { HEADER_HEIGHT } from '../constants'
+import { useNavigation } from '../hooks/useNavigation'
 
 interface NavigationHeaderProps {
   primaryText?: string
