+++ packages/checkout/src/components/SequenceCheckoutProvider/index.ts
@@ -1,0 +1,1 @@
+export * from './SequenceCheckoutProvider'
